feat: “traits” module
- public scope `ScaleFactor`
- and make `scale_factor` always return a Quantity
- traits docs, changelog, whatsnew
