+++ astropy/cosmology/traits.py
@@ -1,0 +1,8 @@
+"""Traits for building Astropy :class:`~astropy.cosmology.Cosmology` objects."""
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+__all__ = [
+    "ScaleFactor",
+]
+
+from ._src.traits import ScaleFactor
