+++ docs/changes/cosmology/17702.api.rst
@@ -1,0 +1,2 @@
+A new public module, ``astropy.cosmology.traits``, has been added to provide building
+blocks for custom cosmologies.
